refactor(server): /api/stats + /api/metrics 라우트 분할 (Plan C #C9 5단계)
- src/server/routes/stats.ts 신규: 6개 read-only 라우트 캡슐화
  (stats, stats/costs, stats/projects, metrics/failure-reasons/throughput/success-rate)
- safeError 적용 (6회)
- dashboard-api.ts에서 11개 unused import 정리 (Get*QuerySchema 5개 + getJobStats 등 6개)
- dashboard-api.ts 1480 → 1345줄 (-135)

남은 도메인: config/setup/repositories/health/doctor/sse/skip-events/auth/health/new-issue
